Custom element <wpvs-detect-screen-size>
This intrdocues a new custom element called <wpvs-detect-screen-size>
which can be placed somewhere in the HTML code to detect the current
screen size. The element contains a dataset-screen-size attribute,
which always contains one of the following values, based upon width
of the viewport:

 * phone
 * tablet
 * screen
 * hires

Additionaly the element raises a screen-size-changed event when the
new viewport width exceeds one of the breakpoints. The event's detail
data then carries a "screen_size" attribute with the new size category.