Custom elements <wpvs-container> and <wpvs-material-card>
This commit brings the element <wpvs-container> which simply positions
its content side-by-side on large screen and below each other on small
screens.

Additionally the element <wpvs-material-card> to be used inside such a
container is included. This elements is the main building block for the
OER material catalog, as it describes a published material:

    <wpvs-material-card
        data-type = "Slides and Script"
        data-name = "Elementary Web Development"
    >
        <material-meta
            data-label = "Format"
            data-value = "HTML"
        >
        </material-meta>
        …
        <material-link
            data-icon  = "icon-globe"
            data-label = "View Online"
            data-href  = "https://…"
        >
        </material-link>
        …
    </wpvs-material-card>