Custom element <wpvs-router>
Custom element <wpvs-router> to integrate the Navigo single page router.
This is just a bare-bones implementation with just the very few features
needed by this side. Namely to change the page content upon navigation to
another url. For this the router works with HTML templates like so:

    <wpvs-router>
        <script type="text/html" data-route="/webdev">
            <wpvs-page data-src="webdev.html" data-title="…"></wpvs-page>
        </script>
        <script type="text/html" data-route="/distsys">
           <wpvs-page data-src="distsys.html" data-title="…"></wpvs-page>
        </script>
    </wpvs-router>

The charm of the <script> elements here is that the browser doesn't interpret
their content. Therefor the <wpvs-page> elements will only be instantiated
when they are picked up by the router to be placed on the site. For the router
to know, when to use which template, the URL pattern given in `data-route`
will be used. This can be any valid pattern that is understood by Navigo.