Overall cleanup and stylesheet fix for <wpvs-page>
<wpvs-page> uses an async version of _render() to fetch the page content
and render it. Since this was not correctly handled by the base-class
CustomElement this lead to the global stylesheets not being applied to
the shadow root of <wpvs-page>. Now the parent render() method is also
async and uses await to call the sub-classed _render() method, so that
the stylesheets will be correctly applied.

Also some redundant commented code has been removed from the other
classes.