Custom element <wpvs-image> and content for the imprint
In Germany you usually need an imprint if you publish a somewhat
commercial looking website. So better be safe than sorry and also have
one. To make it look nice it also contains a few images, which need to
behave responsively. This is done with the new <wpvs-image> element,
like this:

  <wpvs-image data-src="img.jpg" data-alt="…" data-height="15em">
  </wpvs-image>